feat: Set router to handle pages - 프로젝트 진행을 위해 라우터 구조 설정. - 파일 업로드 하기 전(인트로 페이지), 파일 업로드 한 후(대시보드   페이지)로 페이지 구성함.
