+++ src/App.js
@@ -1,24 +1,20 @@
-import React, {useEffect, useState} from 'react';
-import './App.css'; 
-import axios from 'axios';
+import Intro from './component/Intro';
+import Dashboard from './component/Dashboard';
+import {BrowserRouter,Route, Switch} from 'react-router-dom';
 
-  function App() {
-      const [history, setHistory] = useState([]);
-
-      useEffect(() => {
-        fetchChatHistory();
-      }, []);
-    
-      const fetchChatHistory = async () => {
-        const res = await axios('https://programming.coffee/history');
-        setHistory(res.data);
-      };
-      
-    return(
-      <div>
-        {history.map((history, index) => <div key={index}>[{history.date}] {history.user}: {history.message}</div>)}
-      </div>
-    );
+export default function App(){
+  return (
+    <div>
+      <BrowserRouter>
+        <Switch>
+          <Route exact path ="/">
+            <Intro />
+          </Route>
+          <Route path="/info">
+            <Dashboard />
+          </Route>
+        </Switch>
+      </BrowserRouter>
+    </div>
+  );
 }
-
-export default App;
